2.0

+++ vitomuci.js
@@ -41,6 +41,7 @@ async function vitomuci(dir, op, process) {
         cover: false,
         rename: false,
         metadata: false,
+        full: false,
     }, op);
 
     //parse time stamps to seconds
@@ -88,11 +89,17 @@ async function vitomuci(dir, op, process) {
 
     }
     //get files
+    const fileSpinner = ora(`searching for files...`).start();
     let files = getFiles(directory);
     //check if files are media files
     files = verifyFiles(files);
+    if (typeof files === "undefined" || files.length == 0) {
+        fileSpinner.fail("no file where found inside " + directory)
+        throw "no files where found inside " + directory;
+    }
+    fileSpinner.succeed(`found ${chalk.blue(files.length)} file(s), start converting...`)
+
     //rename files
-    if (typeof files === "undefined" || files.length == 0) throw "no files where found inside " + directory;
     files = options.rename ? rename(files) : files;
     let baseDirectory = path.dirname(files[0]);
     let outputDirectory = path.join(baseDirectory, "audio");
@@ -101,38 +108,34 @@ async function vitomuci(dir, op, process) {
     if (!fs.existsSync(outputDirectory))
         fs.mkdirSync(outputDirectory);
 
-
-    console.log(`found ${chalk.blue(files.length)} file(s), start converting...`);
-
-    //main loop
+    //Split track
     for (let item of files) {
         let seconds = await getFileLength(item);
-        await convertToMp3(baseDirectory, item);
         let filename = path.basename(item);
-        let removeType = filename.substr(0, filename.lastIndexOf(".")) || filename;
         await splitTrack(baseDirectory, outputDirectory, filename, Number(seconds));
-        await deleteFile(path.join(baseDirectory, "temp.mp3"));
-
     }
-
-    let coverPath = await getCoverPicture(files[0], baseDirectory, options.startAt);
 
     //set metadata name to first file in array if not set
     if (options.name === "") {
         let filename = path.basename(files[0])
         options.name = filename.substr(0, filename.lastIndexOf(".")) || filename;
     }
+    //take cover picture
+    let cover = "";
+    if (options.cover)
+        coverPath = await getCoverPicture(files[0], baseDirectory, options.startAt);
 
     //updating meta data
     if (options.metadata) {
+        const metadataSpinner = ora(`searching for files...`).start();
         files = fs.readdirSync(outputDirectory);
         for (let file of files) {
             await writeMusicMetadata(path.join(outputDirectory, file), options.name, coverPath);
         }
-        console.log(`updated metadata of ${chalk.blue(files.length)} file(s)`);
+        metadataSpinner.succeed(`updated metadata of ${chalk.blue(files.length)} file(s)`);
     }
 
-    await deleteFile(coverPath);
+    if (options.cover) await deleteFile(coverPath);
 
 }
 
@@ -142,13 +145,13 @@ async function vitomuci(dir, op, process) {
  * packages that require it
  */
 function checkffmpeg() {
-
     ffmpeg.setFfmpegPath(ffmpegPath);
     ffmpeg.setFfprobePath(ffprobePath);
     process.env.FFMPEG_PATH = ffmpegPath;
     ffprobe.FFPROBE_PATH = ffprobePath;
     ffmetadata = require("ffmetadata");
-    console.log(chalk.green("ffmpeg installed at:" + ffmpegPath));
+    console.log(chalk.grey("ffmpeg installed at:" + ffmpegPath));
+    return ffmpegPath;
 }
 
 
@@ -162,6 +165,7 @@ function checkffmpeg() {
  */
 function getFiles(input) {
     try {
+
         //cli supports regex matching
         if (!isUrl(processArgv[2]) && fileExists.sync(directory)) {
             let files = [];
@@ -178,11 +182,10 @@ function getFiles(input) {
         }
         //directory
         if (fs.lstatSync(input).isDirectory()) {
-            console.log("searching " + chalk.blue(input) + " for files...");
             let files = [];
             fs.readdirSync(input).forEach(file => {
                 let stats = fs.statSync(path.join(input, file));
-                if (stats.isFile() && !(file === "temp.mp3"))
+                if (stats.isFile())
                     files.push(path.join(input, file));
             });
             return (files.sort());
@@ -190,7 +193,6 @@ function getFiles(input) {
         throw "no dir"
     } catch (error) {
         //falls back here if cli doesnt supports regex matching
-        console.log("searching for matching file(s)... " + input);
         //remove brackets
         let removeB = "";
         for (let i = 0; i < input.length; i++) {
@@ -224,27 +226,6 @@ function verifyFiles(files) {
 
 
 /**
- * Converts a media file into a mp3 file called temp.mp3
- * @param {*} baseDirectory 
- * @param {*} input 
- */
-function convertToMp3(baseDirectory, input) {
-    return new Promise((resolve, reject) => {
-        let fileInfo;
-        const spinner = ora(`converting ${input} to mp3`).start();
-        ffmpeg(input).format("mp3").save(baseDirectory + "/temp.mp3").on("error", console.error)
-            .on("codecData", function (data) {
-                fileInfo = data;
-            }).on("progress", function (progress) {
-                spinner.text = `converting ${input} to mp3: ${chalk.blue(progress.timemark)}`;
-            }).on("end", function (stdout, stderr) {
-                spinner.succeed(`converting ${input} to mp3`);
-                resolve(fileInfo);
-            });
-    });
-};
-
-/**
  * Extracts one clip out of a longer mp3 file using the 
  * seekInput and duration fuction.
  * Gets called when splitting up a larger file smaller ones
@@ -258,13 +239,15 @@ function segmentMp3(input, output, start, duration) {
         ffmpeg(input).seekInput(start).duration(duration).save(output).on("error", console.error)
             .on("end", function (stdout, stderr) {
                 resolve();
+            }).on('error', function (err, stdout, stderr) {
+                reject('Cannot process video: ' + err.message);
             });
     });
 };
 
 
 /**
- * Splits a mp3 file into multiple smaler sized parts and renames them
+ * Splits a mp3 file into multiple smaller sized parts and renames them
  * if part is shorter than 30 seconds it gets skipped
  * @param {String} baseDirectory 
  * @param {String} outputDirectory 
@@ -272,21 +255,32 @@ function segmentMp3(input, output, start, duration) {
  * @param {Number} duration 
  */
 async function splitTrack(baseDirectory, outputDirectory, name, duration) {
-    let durationIndex = options.startAt;
     let parts = 0;
     const spinner = ora(`splitting ${name} into ${chalk.blue(parts + 1)} parts`).start();
+
+    //if you dont want seprate clips
+    if (options.full) {
+        let ext = path.extname(name);
+        let newName = path.removeExt(name, ext);
+        await segmentMp3(path.join(baseDirectory, name), path.join(outputDirectory, newName + ".mp3"), 0, duration);
+        spinner.succeed(`Converted ${chalk.blue(newName)} into mp3`);
+        return;
+    }
+
+    let durationIndex = options.startAt;
+
     while ((durationIndex + options.duration) <= (duration - options.endAt)) {
         spinner.text = `splitting ${name} into ${chalk.blue(parts + 1)} parts`;
-        await segmentMp3(path.join(baseDirectory, "temp.mp3"), path.join(outputDirectory, getSegmentName(name, durationIndex, durationIndex + options.duration)), durationIndex, options.duration);
+        await segmentMp3(path.join(baseDirectory, name), path.join(outputDirectory, getSegmentName(name, durationIndex, durationIndex + options.duration)), durationIndex, options.duration);
         durationIndex += options.duration;
         parts++;
     }
     if (((duration - options.endAt) - durationIndex) >= 30) {
         spinner.text = `splitting ${name} into ${chalk.blue(parts + 1)} parts`;
-        await segmentMp3(path.join(baseDirectory, "temp.mp3"), path.join(outputDirectory, getSegmentName(name, durationIndex, duration - options.endAt)), durationIndex, options.duration);
+        await segmentMp3(path.join(baseDirectory, name), path.join(outputDirectory, getSegmentName(name, durationIndex, duration - options.endAt)), durationIndex, options.duration);
         parts++;
     }
-    spinner.succeed(`Splitted ${name} into ${chalk.blue(parts)} parts`);
+    spinner.succeed(`splitted ${name} into ${chalk.blue(parts)} parts`);
 
 }
 
@@ -341,6 +335,7 @@ function stringToSeconds(timeString) {
 function getFileLength(file) {
     return new Promise((resolve, reject) => {
         ffprobe(file, (err, probeData) => {
+            if(err) reject(err);
             resolve(probeData.format.duration);
         });
     });
@@ -386,8 +381,7 @@ function writeMusicMetadata(file, compilationName, cover) {
  * @param {String} picTime 
  */
 function getCoverPicture(file, baseDirectory, picTime) {
-    if (options.cover)
-        console.log(`took cover picture from ${chalk.blue(file)} at ${chalk.blue(picTime)}`);
+    const coverPicture = ora(`took cover picture from ${chalk.blue(file)} at ${chalk.blue(picTime)}`).start();
     return new Promise((resolve, reject) => {
         ffmpeg(file)
             .screenshots({
@@ -395,8 +389,11 @@ function getCoverPicture(file, baseDirectory, picTime) {
                 filename: path.join(baseDirectory, "cover.jpg"),
                 size: "320x240"
             }).on("end", function (stdout, stderr) {
+                coverPicture.succeed(`took cover picture from ${chalk.blue(file)} at ${chalk.blue(picTime)}`);
                 resolve(path.join(baseDirectory, "cover.jpg"));
-            });
+            }).on('error', function (err, stdout, stderr) {
+                coverPicture.fail(chalk.red(err.message));
+            });;
     });
 };
 
@@ -409,7 +406,7 @@ function deleteFile(file) {
     return new Promise((resolve, reject) => {
         fs.unlink(file, function (error) {
             if (error) {
-                throw error;
+                reject(error);
             }
             resolve();
         });
@@ -423,6 +420,8 @@ function deleteFile(file) {
  * @param {Array} files 
  */
 function rename(files) {
+    const spinner = ora(`renaming files...`).start();
+
     let renamedFiles = [];
     files.forEach(function (file) {
         let basename = path.basename(file);
@@ -433,7 +432,7 @@ function rename(files) {
         renamedFiles.push(newName);
         fs.renameSync(file, newName);
     });
-    console.log("renamed files to " + renamedFiles);
+    spinner.succeed(`renamed ${chalk.blue(renamedFiles.length)} files.`);
     return renamedFiles;
 }
